debug logs

+++ cpp/src/drpt/io/file_reader.cpp
@@ -471,7 +471,7 @@ drpt::ImageReader::mpi_file_read (string path, int rank, int world_size, int ove
           if (x > 1e+05)
              x =0;
 //          float x = (float) (c);
-          std::srand(static_cast<unsigned float>(std::time(nullptr)));
+          std::srand(static_cast<float>(std::time(nullptr)));
 
           // Generate a random number between 0 and 255
           float randomNumber = std::rand() % 256;
